Add comprehensive stress test suite for kernel primitives
- Add stress_test.c/h with 19 stress test tasks
- Stress semaphores: fast hammer, medium contention, mutex patterns
- Stress pipes: flood test, multi-producer, variable size messages
- Stress scheduler: yield spammer, sleep patterns, CPU hog
- Add verification: sequence ordering, data integrity, error tracking
- Display real-time stats with PASS/FAIL verification status
- Fix display box drawing to use ASCII for terminal compatibility
- Fix pointer-to-int cast warning for 64-bit host builds
- Increase MAX_TASKS to 48 to accommodate stress tests

+++ Core/Src/kernel/task.c
@@ -381,11 +381,12 @@ void os_create_task(task_t *task, void (*function)(void), uint32_t *stack, uint3
     uint32_t *stack_top = stack + stack_size - 1;
     
     // Push in reverse order (high to low address)
+    // Cast through uintptr_t to avoid warnings on 64-bit hosts during testing
 
-    *(stack_top--) = 0x01000000;         // PSR
-    *(stack_top--) = (uint32_t)function; // PC
+    *(stack_top--) = 0x01000000;                      // PSR
+    *(stack_top--) = (uint32_t)(uintptr_t)function;   // PC
 
-    *(stack_top--) = (uint32_t)os_exit_task;         // LR
+    *(stack_top--) = (uint32_t)(uintptr_t)os_exit_task;  // LR
     *(stack_top--) = 0;                  // R12
     *(stack_top--) = 0;                  // R3
     *(stack_top--) = 0;                  // R2
